Add solution to Day 1 Part 2
I didn't realize there was a part 2 until I noticed Day 2 Part 2. I
didn't bother with automated tests for this one--just ran the test cases
manually on the cli like I did for part 1--because I just sort of wanted
to get that old work done and move on to TDDing the solutions properly
for the rest of the days.